+++ src/Webiny/Component/Entity/Attribute/DateAttributeAbstract.php
@@ -8,7 +8,7 @@
 namespace Webiny\Component\Entity\Attribute;
 
 use MongoDB\BSON\UTCDatetime;
-use Webiny\Component\Entity\Validation\ValidationException;
+use Webiny\Component\Entity\EntityValidationException;
 use Webiny\Component\StdLib\StdObject\DateTimeObject\DateTimeObject;
 use Webiny\Component\StdLib\StdObject\DateTimeObject\DateTimeObjectException;
 
@@ -100,7 +100,7 @@ abstract class DateAttributeAbstract extends AttributeAbstract
      *
      * @param $value
      *
-     * @throws ValidationException
+     * @throws EntityValidationException
      * @return $this
      */
     protected function validate(&$value)