+++ src/Webiny/Component/Entity/Attribute/IntegerAttribute.php
@@ -7,7 +7,7 @@
 
 namespace Webiny\Component\Entity\Attribute;
 
-use Webiny\Component\Entity\Validation\ValidationException;
+use Webiny\Component\Entity\EntityValidationException;
 
 /**
  * IntegerAttribute
@@ -42,7 +42,7 @@ class IntegerAttribute extends AttributeAbstract
      *
      * @param $value
      *
-     * @throws ValidationException
+     * @throws EntityValidationException
      * @return $this
      */
     protected function validate(&$value)