+++ src/Webiny/Component/Entity/Validators/CountryCode.php
@@ -1,0 +1,27 @@
+<?php
+namespace Webiny\Component\Entity\Validators;
+
+use Webiny\Component\Entity\Attribute\AttributeAbstract;
+use Webiny\Component\Entity\EntityValidatorInterface;
+use Webiny\Component\Validation\ValidationTrait;
+
+class CountryCode implements EntityValidatorInterface
+{
+    use ValidationTrait;
+
+    /**
+     * @inheritDoc
+     */
+    public function getName()
+    {
+        return 'countryCode';
+    }
+    
+    /**
+     * @inheritDoc
+     */
+    public function validate(AttributeAbstract $attribute, $data, $params = [])
+    {
+        return $this->validation()->validate($data, 'countryCode');
+    }
+}