+++ src/Webiny/Component/Entity/Validators/Url.php
@@ -1,0 +1,27 @@
+<?php
+namespace Webiny\Component\Entity\Validators;
+
+use Webiny\Component\Entity\Attribute\AttributeAbstract;
+use Webiny\Component\Entity\EntityValidatorInterface;
+use Webiny\Component\Validation\ValidationTrait;
+
+class Url implements EntityValidatorInterface
+{
+    use ValidationTrait;
+
+    /**
+     * @inheritDoc
+     */
+    public function getName()
+    {
+        return 'url';
+    }
+
+    /**
+     * @inheritDoc
+     */
+    public function validate(AttributeAbstract $attribute, $data, $params = [])
+    {
+        return $this->validation()->validate($data, 'url');
+    }
+}